+++ R/primitive-title.R
@@ -38,6 +38,12 @@
 #'   x.sec = primitive_title("Horizontal Title"),
 #'   y.sec = primitive_title(c("along vertical", "Multiple tiles"))
 #' )
+#'
+#' # 'Real' titles occur once per plot.
+#' # Primitive titles repeat over facets and hide the 'real' title.
+#' p + facet_wrap(~ drv) +
+#'   guides(x = primitive_title("I am a repeated subtitle")) +
+#'   labs(x = "I am the hidden real title")
 primitive_title = function(title = waiver(), angle = waiver(),
                            theme = NULL, position = waiver()) {
   if (!is_waive(angle)) {
@@ -78,7 +84,8 @@ PrimitiveTitle <- ggproto(
 
   extract_key = function(scale, aesthetic, ...) {
     # Need to keep track of limits for r/r.sec positions
-    data_frame0(!!aesthetic := c(-Inf, Inf), .value = scale$get_limits())
+    limits <- range(scale$continuous_range %||% scale$get_limits())
+    data_frame0(!!aesthetic := c(-Inf, Inf), .value = limits)
   },
 
   extract_params = function(scale, params, title = waiver(), ...) {
@@ -216,8 +223,6 @@ draw_cart_title <- function(label, elements, params) {
 
   title    <- elements$title
   position <- params$position
-  hjust <- title$hjust
-  vjust <- title$vjust
 
   angle <- (params$angle %|W|% NULL) %||% title$angle
 
@@ -227,6 +232,10 @@ draw_cart_title <- function(label, elements, params) {
   })
   widths  <- width_cm(singles)
   heights <- height_cm(singles)
+
+  justs <- rotate_just(element = title)
+  hjust <- justs$hjust
+  vjust <- justs$vjust
 
   if (position %in% c("left", "right")) {
     x <- hjust