+++ R/key-.R
@@ -40,7 +40,7 @@
 #'   transformation to calculate positions. It is only advisable to set the
 #'   `prescale_base` argument when the data have already been log-transformed.
 #'   When using a log-transform in the scale or in
-#'   [`coord_trans()`][ggplot2::coord_trans], the default `NULL` is recommended.
+#'   [`coord_transform()`][ggplot2::coord_transform], the default `NULL` is recommended.
 #' @param negative_small A `<numeric[1]>` setting the smallest absolute value
 #'   that is marked with a tick in case the scale limits include 0 or negative
 #'   numbers.