+++ NEWS.md
@@ -1,3 +1,5 @@
+# legendry (development version)
+
 # legendry 0.2.4
 
 This is a patch release fixing a few bugs.