+++ NEWS.md
@@ -1,6 +1,9 @@
 # legendry (development version)
 
 * `compose_stack(side.titles)` also work in non-position guides (#101)
+* Updated title mechanism to comply with ggplot2#6200 (#104)
+* `key_group_split()` tries to preserve order of labels better, rather than 
+  re-sorting them (#100)
 
 # legendry 0.2.4