+++ NEWS.md
@@ -1,5 +1,6 @@
 # legendry (development version)
 
+* `compose_stack(side.titles)` also work in non-position guides (#101)
 * Updated title mechanism to comply with ggplot2#6200 (#104)
 * `key_group_split()` tries to preserve order of labels better, rather than 
   re-sorting them (#100)