+++ NEWS.md
@@ -2,6 +2,10 @@
 
 * New arguments `guide_axis_base(subtitle)` and `guide_axis_nested(subtitle)` 
   make it easier to title custom axes (#102)
+* `compose_stack(side.titles)` also work in non-position guides (#101)
+* Updated title mechanism to comply with ggplot2#6200 (#104)
+* `key_group_split()` tries to preserve order of labels better, rather than 
+  re-sorting them (#100)
 
 # legendry 0.2.4